+++ beets/plugins.py
@@ -54,6 +54,12 @@ class BeetsPlugin(object):
         commands that should be added to beets' CLI.
         """
         return ()
+    
+    def queries(self):
+        """Should return a dict mapping prefixes to PluginQuery
+        subclasses.
+        """
+        return {}
 
     def track_distance(self, item, info):
         """Should return a (distance, distance_max) pair to be added
@@ -92,6 +98,7 @@ class BeetsPlugin(object):
         database schema is not (currently) extended.
         """
         return {}
+
 
     listeners = None
 
@@ -207,6 +214,15 @@ def commands():
     out = []
     for plugin in find_plugins():
         out += plugin.commands()
+    return out
+
+def queries():
+    """Returns a dict mapping prefix strings to beet.library.PluginQuery
+    subclasses all loaded plugins.
+    """
+    out = {}
+    for plugin in find_plugins():
+        out.update(plugin.queries())
     return out
 
 def track_distance(item, info):