+++ test/test_query.py
@@ -20,82 +20,64 @@ import _common
 from _common import unittest
 import beets.library
 
-pqp = beets.library.CollectionQuery._parse_query_part
+pqp = beets.library.parse_query_part
 
 some_item = _common.item()
 
 class QueryParseTest(unittest.TestCase):
     def test_one_basic_term(self):
         q = 'test'
-        r = (None, 'test', False)
+        r = (None, 'test', beets.library.SubstringQuery)
         self.assertEqual(pqp(q), r)
 
     def test_one_keyed_term(self):
         q = 'test:val'
-        r = ('test', 'val', False)
+        r = ('test', 'val', beets.library.SubstringQuery)
         self.assertEqual(pqp(q), r)
 
     def test_colon_at_end(self):
         q = 'test:'
-        r = (None, 'test:', False)
+        r = (None, 'test:', beets.library.SubstringQuery)
         self.assertEqual(pqp(q), r)
 
     def test_one_basic_regexp(self):
         q = r':regexp'
-        r = (None, 'regexp', True)
+        r = (None, 'regexp', beets.library.RegexpQuery)
         self.assertEqual(pqp(q), r)
 
     def test_keyed_regexp(self):
         q = r'test::regexp'
-        r = ('test', 'regexp', True)
+        r = ('test', 'regexp', beets.library.RegexpQuery)
         self.assertEqual(pqp(q), r)
 
     def test_escaped_colon(self):
         q = r'test\:val'
-        r = (None, 'test:val', False)
+        r = (None, 'test:val', beets.library.SubstringQuery)
         self.assertEqual(pqp(q), r)
 
     def test_escaped_colon_in_regexp(self):
         q = r':test\:regexp'
-        r = (None, 'test:regexp', True)
+        r = (None, 'test:regexp', beets.library.RegexpQuery)
         self.assertEqual(pqp(q), r)
 
-class AnySubstringQueryTest(unittest.TestCase):
+class AnyFieldQueryTest(unittest.TestCase):
     def setUp(self):
         self.lib = beets.library.Library(':memory:')
         self.lib.add(some_item)
 
     def test_no_restriction(self):
-        q = beets.library.AnySubstringQuery('title')
+        q = beets.library.AnyFieldQuery('title', beets.library.ITEM_KEYS,
+                                        beets.library.SubstringQuery)
         self.assertEqual(self.lib.items(q).next().title, 'the title')
 
     def test_restriction_completeness(self):
-        q = beets.library.AnySubstringQuery('title', ['title'])
+        q = beets.library.AnyFieldQuery('title', ['title'],
+                                        beets.library.SubstringQuery)
         self.assertEqual(self.lib.items(q).next().title, 'the title')
 
     def test_restriction_soundness(self):
-        q = beets.library.AnySubstringQuery('title', ['artist'])
-        self.assertRaises(StopIteration, self.lib.items(q).next)
-
-class AnyRegexpQueryTest(unittest.TestCase):
-    def setUp(self):
-        self.lib = beets.library.Library(':memory:')
-        self.lib.add(some_item)
-
-    def test_no_restriction(self):
-        q = beets.library.AnyRegexpQuery(r'^the ti')
-        self.assertEqual(self.lib.items(q).next().title, 'the title')
-
-    def test_restriction_completeness(self):
-        q = beets.library.AnyRegexpQuery(r'^the ti', ['title'])
-        self.assertEqual(self.lib.items(q).next().title, 'the title')
-
-    def test_restriction_soundness(self):
-        q = beets.library.AnyRegexpQuery(r'^the ti', ['artist'])
-        self.assertRaises(StopIteration, self.lib.items(q).next)
-
-    def test_restriction_soundness_2(self):
-        q = beets.library.AnyRegexpQuery(r'the ti$', ['title'])
+        q = beets.library.AnyFieldQuery('title', ['artist'],
+                                        beets.library.SubstringQuery)
         self.assertRaises(StopIteration, self.lib.items(q).next)
 
 # Convenient asserts for matching items.